object_ai/pyproject.toml
[tool.poetry]
name = "object-ai"
version = "0.1.0"
description = ""
authors = ["Kevin Fortier <kevin.r.fortier@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8, <4"
Flask = "^2.1.2"
python-dotenv = "^0.20.0"
torch = "^1.12.0"
torchvision = "^0.13.0"
pycocotools = "^2.0.4"
awscli = "^1.25.22"
labelbox = "^3.23.3"
prefect = ">=2.0b"
greenlet = "^1.1.2"
boto3 = "^1.24.23"
geojson = "^2.5.0"
opencv-python = "^4.6.0"
Shapely = "^1.8.2"
pyproj = "^3.3.1"
pyGeoTile = "^1.0.6"
imagesize = "^1.4.1"
black = "^22.6.0"

[tool.poetry.dev-dependencies]
pytest = "^3.4"